requirements.txt
datasets>=2.14.0
pandas>=2.0.0
torch
transformers>=4.40.0
huggingface_hub
accelerate
safetensors
wandb
python-dotenv
trl>=0.8.0
matplotlib
vllm
